ENH: Remove the `ParabolicMorphology.remote.cmake` file.
Remove the `ParabolicMorphology.remote.cmake`. The file is only useful
when being used in ITK or other source code calling the remote.

The file is already in the ITK remote modules, and hence the module is
available as a remote in ITK.

Also, this helps to keep the root of the repository clean.